Refactor Session as a Union type
- Keeping separate session and games entries in the Model would allow
for impossible states, since I could end up with games and no token, or
a valid session with an empty string as a token.
- The union type allows me to encapsulate a user’s session as a variant
vs not being logged in. I’m still dealing with Maybes so there’s likely
some holes there but I feel infinitely better about this representation.